+++ app/src/main/java/com/forhopssake/xbrew/data/Malt.java
@@ -2,6 +2,6 @@ package com.forhopssake.xbrew.data;
 
 public interface Malt extends Ingredient {
     public String getSupplier();
-    public double getPotential();
-    public double getColour();
+    public float getPotential();
+    public float getColour();
 }
