+++ app/src/main/java/com/forhopssake/xbrew/data/impl/Fermentable.java
@@ -7,12 +7,12 @@ public class Fermentable {
     private int version = 1; // Required
     private FermentableType type; // Required
     // Weight of the fermentable, extract or sugar in Kilograms.
-    private double amount; // Required
+    private float amount; // Required
     // Percent dry yield (fine grain) for the grain, or the raw
     // yield by weight if this is an extract adjunct or sugar.
-    private double yield; // Required
+    private float yield; // Required
     // The color of the item in Lovibond Units (SRM for liquid extracts).
-    private double colour; // Required
+    private float colour; // Required
     // May be TRUE if this item is normally added after the boil.
     // The default value is FALSE since most grains are added during the mash or boil.
     private boolean addAfterBoil;
@@ -23,19 +23,19 @@ public class Fermentable {
 
     // Percent difference between the coarse grain yield and fine grain yield.
     // Only appropriate for a "Grain" or "Adjunct" type, otherwise this value is ignored.
-    private double courseFineFine;
+    private float courseFineFine;
     // Percent moisture in the grain.
     // Only appropriate for a "Grain" or "Adjunct" type, otherwise this value is ignored.
-    private double moisture;
+    private float moisture;
     // The diastatic power of the grain as measured in "Lintner" units.
     // Only appropriate for a "Grain" or "Adjunct" type, otherwise this value is ignored.
-    private double diasticPower;
+    private float diasticPower;
     // The percent protein in the grain.
     // Only appropriate for a "Grain" or "Adjunct" type, otherwise this value is ignored.
-    private double protein;
+    private float protein;
     // The recommended maximum percentage (by weight)
     // this ingredient should represent in a batch of beer.
-    private double maxInBatch;
+    private float maxInBatch;
     // TRUE if it is recommended the grain be mashed, FALSE if it can be steeped.
     // A value of TRUE is only appropriate for a "Grain" or "Adjunct" types.
     // The default value is FALSE.  Note that this does NOT indicate whether the grain
@@ -73,27 +73,27 @@ public class Fermentable {
         this.type = type;
     }
 
-    public double getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
 
-    public double getYield() {
+    public float getYield() {
         return yield;
     }
 
-    public void setYield(double yield) {
+    public void setYield(float yield) {
         this.yield = yield;
     }
 
-    public double getColour() {
+    public float getColour() {
         return colour;
     }
 
-    public void setColour(double colour) {
+    public void setColour(float colour) {
         this.colour = colour;
     }
 
@@ -121,43 +121,43 @@ public class Fermentable {
         this.supplier = supplier;
     }
 
-    public double getCourseFineFine() {
+    public float getCourseFineFine() {
         return courseFineFine;
     }
 
-    public void setCourseFineFine(double courseFineFine) {
+    public void setCourseFineFine(float courseFineFine) {
         this.courseFineFine = courseFineFine;
     }
 
-    public double getMoisture() {
+    public float getMoisture() {
         return moisture;
     }
 
-    public void setMoisture(double moisture) {
+    public void setMoisture(float moisture) {
         this.moisture = moisture;
     }
 
-    public double getDiasticPower() {
+    public float getDiasticPower() {
         return diasticPower;
     }
 
-    public void setDiasticPower(double diasticPower) {
+    public void setDiasticPower(float diasticPower) {
         this.diasticPower = diasticPower;
     }
 
-    public double getProtein() {
+    public float getProtein() {
         return protein;
     }
 
-    public void setProtein(double protein) {
+    public void setProtein(float protein) {
         this.protein = protein;
     }
 
-    public double getMaxInBatch() {
+    public float getMaxInBatch() {
         return maxInBatch;
     }
 
-    public void setMaxInBatch(double maxInBatch) {
+    public void setMaxInBatch(float maxInBatch) {
         this.maxInBatch = maxInBatch;
     }
